+++ src/main/java/org/ewgf/controllers/StatisticsController.java
@@ -4,41 +4,33 @@ import jakarta.servlet.http.HttpServletRequest;
 
 import lombok.extern.slf4j.Slf4j;
 
-import org.ewgf.utils.TekkenDataMapper;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
 
-import org.ewgf.interfaces.*;
 import org.ewgf.dtos.*;
 import org.ewgf.models.TekkenStatsSummary;
-import org.ewgf.repositories.AggregatedStatisticsRepository;
 import org.ewgf.repositories.TekkenStatsSummaryRepository;
-import org.ewgf.services.CharacterAnalyticsService;
+import org.ewgf.services.StatisticsService;
 
 import java.util.*;
-import java.util.stream.Collectors;
 
 @RestController
 @RequestMapping("/statistics")
 @Slf4j
 public class StatisticsController {
     private final TekkenStatsSummaryRepository tekkenStatsSummaryRepository;
-    private final AggregatedStatisticsRepository aggregatedStatisticsRepository;
-    private final CharacterAnalyticsService characterAnalyticsService;
+    private final StatisticsService statisticsService;
 
     public StatisticsController(
             TekkenStatsSummaryRepository tekkenStatsSummaryRepository,
-            AggregatedStatisticsRepository aggregatedStatisticsRepository,
-            CharacterAnalyticsService characterAnalyticsService)
+            StatisticsService statisticsService)
     {
         this.tekkenStatsSummaryRepository = tekkenStatsSummaryRepository;
-        this.aggregatedStatisticsRepository = aggregatedStatisticsRepository;
-        this.characterAnalyticsService = characterAnalyticsService;
+        this.statisticsService = statisticsService;
     }
 
     @GetMapping("/stats-summary")
-    public ResponseEntity<TekkenStatsSummaryDTO> getPlayerCount(HttpServletRequest request) throws InterruptedException
-    {
+    public ResponseEntity<TekkenStatsSummaryDTO> getPlayerCount(HttpServletRequest request) throws InterruptedException {
         log.info("Received request for stats summary");
         return tekkenStatsSummaryRepository.getTekkenStatsSummary()
                 .map(this::convertToDTO)
@@ -47,62 +39,37 @@ public class StatisticsController {
     }
 
     @GetMapping("/version-popularity")
-    public ResponseEntity<Map<String, CharacterPopularityDTO>> getVersionPopularity()
-    {
+    public ResponseEntity<Map<String, CharacterPopularityDTO>> getVersionPopularity() throws Exception {
         log.info("Fetching popularity stats for all game versions");
-        try {
-            Map<String, CharacterPopularityDTO> popularity = characterAnalyticsService.getAllVersionPopularity();
-            return ResponseEntity.ok(popularity);
-        }
-        catch (Exception e) {
-            log.error("Error fetching version-specific popularity statistics", e);
-            return ResponseEntity.internalServerError().build();
-        }
+        Map<String, CharacterPopularityDTO> popularity = statisticsService.getAllVersionPopularity();
+        return ResponseEntity.ok(popularity);
     }
 
     @GetMapping("/version-winrates")
-    public ResponseEntity<Map<String, CharacterWinratesDTO>> getVersionWinrates()
-    {
+    public ResponseEntity<Map<String, CharacterWinratesDTO>> getVersionWinrates() throws Exception {
         log.info("Fetching winrates for all game versions");
-        try {
-            Map<String, CharacterWinratesDTO> winrates = characterAnalyticsService.getAllVersionWinrates();
-            return ResponseEntity.ok(winrates);
-        }
-        catch (Exception e) {
-            log.error("Error fetching version-specific winrates", e);
-            return ResponseEntity.internalServerError().build();
-        }
+        Map<String, CharacterWinratesDTO> winrates = statisticsService.getAllVersionWinrates();
+        return ResponseEntity.ok(winrates);
     }
 
     @GetMapping("/top-popularity")
-    public ResponseEntity<CharacterPopularityDTO> getTop5CharacterPopularityStats() {
+    public ResponseEntity<CharacterPopularityDTO> getTop5CharacterPopularityStats() throws Exception {
         log.info("Fetching top 5 popular characters");
-        try {
-            CharacterPopularityDTO popularity = characterAnalyticsService.getTopCharacterPopularity();
-            return ResponseEntity.ok(popularity);
-        } catch (Exception e) {
-            log.error("Error fetching character popularity statistics", e);
-            return ResponseEntity.internalServerError().build();
-        }
+        CharacterPopularityDTO popularity = statisticsService.getTopCharacterPopularity();
+        return ResponseEntity.ok(popularity);
     }
 
     @GetMapping("/top-winrates")
-    public ResponseEntity<CharacterWinratesDTO> getTop5CharacterWinratesStats() {
+    public ResponseEntity<CharacterWinratesDTO> getTop5CharacterWinratesStats() throws Exception {
         log.info("Fetching top 5 highest winrate characters");
-        try {
-            CharacterWinratesDTO winratesDTO = characterAnalyticsService.getTopCharacterWinrates();
-            return ResponseEntity.ok(winratesDTO);
-        } catch (Exception e) {
-            log.error("Error fetching character popularity statistics", e);
-            return ResponseEntity.internalServerError().build();
-        }
+        CharacterWinratesDTO winratesDTO = statisticsService.getTopCharacterWinrates();
+        return ResponseEntity.ok(winratesDTO);
     }
 
     @GetMapping("/gameVersions")
-    public ResponseEntity<List<Integer>> getGameVersions(HttpServletRequest request) throws InterruptedException
-    {
-        log.info("Received request for gameVerions");
-        return aggregatedStatisticsRepository.getGameVersions()
+    public ResponseEntity<List<Integer>> getGameVersions(HttpServletRequest request) throws InterruptedException {
+        log.info("Received request for gameVersions");
+        return statisticsService.getGameVersions()
                 .map(ResponseEntity::ok)
                 .orElse(ResponseEntity.notFound().build());
     }
@@ -110,39 +77,15 @@ public class StatisticsController {
     @GetMapping("/rankDistribution")
     public ResponseEntity<Map<Integer, RankDistributionDTO>> getAllRankDistributions() {
         log.info("Fetching rank distribution for all versions");
-        Optional<List<Integer>> gameVersions = aggregatedStatisticsRepository.getGameVersions();
-        List<RankDistributionProjection> distributions = aggregatedStatisticsRepository.getAllRankDistributions(gameVersions.get());
-        Map<Integer, RankDistributionDTO> result = new TreeMap<>(Collections.reverseOrder());
-
-        distributions.forEach(dist -> {
-            result.computeIfAbsent(dist.getGameVersion(), k -> new RankDistributionDTO())
-                    .addDistribution(dist.getCategory(),
-                            new RankDistributionEntry(dist.getRank(), dist.getPercentage()));
-        });
-
+        Map<Integer, RankDistributionDTO> result = statisticsService.getAllRankDistributions();
         return ResponseEntity.ok(result);
     }
 
     @GetMapping("/winrate-changes")
     public ResponseEntity<Map<String, List<RankWinrateChangesDTO>>> getWinrateChanges() {
         log.info("Fetching character winrate changes");
-        try {
-            List<WinrateChangesProjection> projections = aggregatedStatisticsRepository.getWinrateChanges();
-            List<RankWinrateChangesDTO> changes = projections.stream()
-                    .map(proj -> new RankWinrateChangesDTO(
-                            TekkenDataMapper.getCharacterName(proj.getCharacterId()),
-                            proj.getChange(),
-                            proj.getTrend(),
-                            proj.getRankCategory()
-                    ))
-                    .collect(Collectors.toList());
-
-            Map<String, List<RankWinrateChangesDTO>> groupedChanges = RankWinrateChangesDTO.groupByRankCategory(changes);
-            return ResponseEntity.ok(groupedChanges);
-        } catch (Exception e) {
-            log.error("Error fetching character winrate changes", e);
-            return ResponseEntity.internalServerError().build();
-        }
+        Map<String, List<RankWinrateChangesDTO>> groupedChanges = statisticsService.getWinrateChanges();
+        return ResponseEntity.ok(groupedChanges);
     }
 
     private TekkenStatsSummaryDTO convertToDTO(TekkenStatsSummary tekkenStatsSummary) {
