+++ src/main/java/org/ewgf/interfaces/CharacterAnalyticsProjection.java
@@ -2,8 +2,6 @@ package org.ewgf.interfaces;
 
 public interface CharacterAnalyticsProjection {
     String getCharacterId();
-    long getTotalWins();
-    long getTotalLosses();
     double getWinratePercentage();
     long getTotalBattles();
     Integer getGameVersion();
