+++ src/main/java/org/ewgf/interfaces/CharacterWinrateProjection.java
@@ -4,7 +4,6 @@ public interface CharacterWinrateProjection {
     String getCharacterId();
     Long getTotalWins();
     Long getTotalLosses();
-    Double getWinRate();
     Integer getGameVersion();
     String getRankCategory();
     String getRegionId();
