--- src/main/java/org/ewgf/interfaces/PlayerAnalyticsProjection.java
@@ -1,15 +1,0 @@
-package org.ewgf.interfaces;
-
-public interface PlayerAnalyticsProjection {
-    String getPlayerId();
-    String getCharacterId();
-    Integer getDanRank();
-    Integer getWins();
-    Integer getLosses();
-    Integer getRegionId();
-    Integer getAreaId();
-
-    default Integer getTotalPlays() {
-        return (getWins() != null && getLosses() != null) ? getWins() + getLosses() : 0;
-    }
-}
