--- src/main/java/org/ewgf/interfaces/PopularCharacterProjection.java
@@ -1,10 +1,0 @@
-package org.ewgf.interfaces;
-
-public interface PopularCharacterProjection {
-    String getRegionId();
-    String getCharacterId();
-    Long getTotalWins();
-    Long getTotalLosses();
-    Long getTotalBattles();
-    Double getWinratePercentage();
-}
